Refactor test.cpp: Add levelOne() function
This change refactors the test.cpp file by adding a new function called
levelOne(). This function is currently empty, but it is intended to
contain the logic for the first level of the game. Including this new
function will allow for better organization and modularization of the
code. No issues are addressed in this commit.

Note: The commit message assumes that the addition of the 'levelOne()'
function is relevant and desired for the project. If there are
additional reasons or specific requirements for this change, please
provide that information so it can be incorporated into the message.